fix(HD-38): Applied fixes
fixed :
'bindActionCreators' is defined but never used.  (lines of code = 1)
img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.  (lines of code = 1)
'bindActionCreators' is defined but never used.  (lines of code = 1)
'logger' is assigned a value but never used.  (lines of code = 1)