Feature/hd 20 (#24)
* add userProfile component and update the book component

refactor the book component in order to create the library, remove the booklist component and create the userProfile component also for the library

* feature/HD-20: create UserRAting component and Details

Add endpoint file, create Details page and UserRating component

* feature/HD-20: update userRating and UserProfile files

updates for Dolan

* feature/HD-20: finish userRating component

finish the component to display the rate of the current user. need to be update after the db conf

* add userProfile component and update the book component

refactor the book component in order to create the library, remove the booklist component and create the userProfile component also for the library

* feature/HD-20: create UserRAting component and Details

Add endpoint file, create Details page and UserRating component

* feature/HD-20: update userRating and UserProfile files

updates for Dolan

* feature/HD-20: finish userRating component

finish the component to display the rate of the current user. need to be update after the db conf

* feature/HD-20: add library and book component

display all the books owned by the selected user

* feature/HD-20: create book style

create style files in order to obtain a 'clean' library

* feature/HD-20: add image and update style

add a 404 image and make a placeholder in the css files

* feature/HD-20: add user style

update style for the user component

* feature/HD-20: fix axios get

* feature/HD-20: fix some troubles

* feature/HD-20: fix user library

* feature/HD-20: improve codefactor

* feature/HD-20: imporve codefactor

* feature/HD-20: codefactor again

* feature/HD-20: ...codefactor